pyproject.toml
[tool.poetry]
name = "stock-trader"
version = "0.1.0"
description = ""
authors = ["Seran Shanmugathas <seran@live.no>"]
readme = "README.md"
packages = [{include = "stock_trader"}]

[tool.poetry.dependencies]
python = "^3.10"
gym = "^0.26.2"
numpy = "^1.26.2"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
yfinance = "^0.2.33"
pyyaml = "^6.0.1"
gymnasium = "^0.29.1"
torch = "^2.1.2"
termcolor = "^2.4.0"
tensorboard = "^2.15.1"
torchvision = "^0.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run-main = "python3 stock_trader/main.py"